# docs/swagger.yaml
basePath: /
definitions:
...
  handler.AgentChatRequest:
    properties:
      history:
        items:
          $ref: '#/definitions/handler.GeminiMessage'
        type: array
      message:
        example: crie um job para rodar a cada 5 minutos batendo em https://httpbin.org/get
        type: string
    type: object
  handler.AgentChatResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/handler.GeminiMessage'
        type: array
      reply:
        example: "Executando ferramenta createJob... Job criado com ID 4a82 com sucesso!
          \U0001F680"
        type: string
    type: object
...
  handler.GeminiFunctionCall:
    properties:
      args:
        additionalProperties: {}
        type: object
      id:
        type: string
      name:
        type: string
    type: object
  handler.GeminiFunctionResponse:
    properties:
      id:
        type: string
      name:
        type: string
      response:
        additionalProperties: {}
        type: object
    type: object
  handler.GeminiMessage:
    properties:
      parts:
        items:
          $ref: '#/definitions/handler.GeminiPart'
        type: array
      role:
        type: string
    type: object
  handler.GeminiPart:
    properties:
      functionCall:
        $ref: '#/definitions/handler.GeminiFunctionCall'
      functionResponse:
        $ref: '#/definitions/handler.GeminiFunctionResponse'
      text:
        type: string
      thoughtSignature:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API do sistema CronFlow para agendamento e monitoramento de execução
    de tarefas distribuídas em background.
  title: CronFlow API
  version: "1.0"
paths:
...
  /v1/agent/chat:
    post:
      consumes:
      - application/json
      description: Envia uma mensagem para o Agente de IA, permitindo agendar jobs,
        testar cURLs, consultar histórico ou gerenciar tarefas por linguagem natural.
      parameters:
      - description: Histórico e nova mensagem do usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AgentChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Resposta do Agente de IA e histórico atualizado
          schema:
            $ref: '#/definitions/handler.AgentChatResponse'
        "400":
          description: Payload inválido ou sem mensagem
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Não autenticado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erros na API do Gemini/Groq ou durante a execução de ferramentas
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Conversar com o Agente de IA do CronFlow
      tags:
      - AI Agent
...
securityDefinitions:
  ApiKeyAuth:
    description: Tipo "Bearer <sua-api-key>" para se autenticar.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"