# internal/api/handler/agent_handler.go
package handler

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"io"
	"log"
	"net/http"
	"time"

	"github.com/JanGustavo/Cron/internal/api/middleware"
	"github.com/JanGustavo/Cron/internal/config"
	"github.com/JanGustavo/Cron/internal/domain/job"
	"github.com/JanGustavo/Cron/internal/service"
)

type AgentHandler struct {
	jobService *service.JobService
	cfg        *config.Config
}

func NewAgentHandler(jobService *service.JobService, cfg *config.Config) *AgentHandler {
	return &AgentHandler{
		jobService: jobService,
		cfg:        cfg,
	}
}

type AgentChatRequest struct {
	Message string          `json:"message" example:"crie um job para rodar a cada 5 minutos batendo em https://httpbin.org/get"`
	History []GeminiMessage `json:"history"`
}

type AgentChatResponse struct {
	Reply   string          `json:"reply" example:"Executando ferramenta createJob... Job criado com ID 4a82 com sucesso! 🚀"`
	History []GeminiMessage `json:"history"`
}

// Gemini Types
type GeminiMessage struct {
	Role  string       `json:"role"`
	Parts []GeminiPart `json:"parts"`
}

type GeminiPart struct {
	Text             string                  `json:"text,omitempty"`
	FunctionCall     *GeminiFunctionCall     `json:"functionCall,omitempty"`
	FunctionResponse *GeminiFunctionResponse `json:"functionResponse,omitempty"`
	ThoughtSignature string                  `json:"thoughtSignature,omitempty"`
}

type GeminiFunctionCall struct {
	ID   string         `json:"id,omitempty"`
	Name string         `json:"name"`
	Args map[string]any `json:"args"`
}

type GeminiFunctionResponse struct {
	ID       string         `json:"id,omitempty"`
	Name     string         `json:"name"`
	Response map[string]any `json:"response"`
}

type GeminiRequest struct {
	Contents          []GeminiMessage          `json:"contents"`
	SystemInstruction *GeminiSystemInstruction `json:"systemInstruction,omitempty"`
	Tools             []GeminiTool             `json:"tools,omitempty"`
}

type GeminiSystemInstruction struct {
	Parts []GeminiPart `json:"parts"`
}

type GeminiTool struct {
	FunctionDeclarations []GeminiFunctionDeclaration `json:"functionDeclarations"`
}

type GeminiFunctionDeclaration struct {
	Name        string       `json:"name"`
	Description string       `json:"description"`
	Parameters  GeminiSchema `json:"parameters"`
}

type GeminiSchema struct {
	Type        string                  `json:"type"`
	Properties  map[string]GeminiSchema `json:"properties,omitempty"`
	Required    []string                `json:"required,omitempty"`
	Description string                  `json:"description,omitempty"`
}

type GeminiResponse struct {
	Candidates []GeminiCandidate `json:"candidates"`
}

type GeminiCandidate struct {
	Content GeminiMessage `json:"content"`
}

// OpenAI/Groq Fallback Types
type OpenAIMessage struct {
	Role       string            `json:"role"`
	Content    string            `json:"content,omitempty"`
	ToolCalls  []OpenAIToolCall  `json:"tool_calls,omitempty"`
	ToolCallID string            `json:"tool_call_id,omitempty"`
}

type OpenAIToolCall struct {
	ID       string           `json:"id"`
	Type     string           `json:"type"`
	Function OpenAIFunction   `json:"function"`
}

type OpenAIFunction struct {
	Name      string `json:"name"`
	Arguments string `json:"arguments"`
}

type OpenAITool struct {
	Type     string             `json:"type"`
	Function OpenAIFunctionDecl `json:"function"`
}

type OpenAIFunctionDecl struct {
	Name        string       `json:"name"`
	Description string       `json:"description"`
	Parameters  GeminiSchema `json:"parameters"`
}

type OpenAIRequest struct {
	Model    string          `json:"model"`
	Messages []OpenAIMessage `json:"messages"`
	Tools    []OpenAITool    `json:"tools,omitempty"`
}

type OpenAIResponse struct {
	Choices []OpenAIChoice `json:"choices"`
}

type OpenAIChoice struct {
	Message OpenAIMessage `json:"message"`
}

var geminiTools = []GeminiTool{
	{
		FunctionDeclarations: []GeminiFunctionDeclaration{
			{
				Name:        "listJobs",
				Description: "Lista todas as tarefas agendadas (Cron Jobs) ativas ou cadastradas no sistema CronFlow.",
				Parameters: GeminiSchema{
					Type: "OBJECT",
				},
			},
			{
				Name:        "createJob",
				Description: "Cria uma nova tarefa agendada (Cron Job) no CronFlow a partir de parâmetros como nome, URL de destino e agendamento.",
				Parameters: GeminiSchema{
					Type: "OBJECT",
					Properties: map[string]GeminiSchema{
						"name": {
							Type:        "STRING",
							Description: "Nome legível/amigável da tarefa (ex: Sincronização de Pedidos)",
						},
						"schedule": {
							Type:        "STRING",
							Description: "Agendamento cron padrão (ex: '*/5 * * * *') ou intervalo simplificado (ex: 'every:1h', 'every:30m'). Traduza pedidos informais do usuário para o formato de agendamento do CronFlow.",
						},
						"url": {
							Type:        "STRING",
							Description: "URL HTTP de destino (webhook) que receberá o disparo agendado (ex: https://meu-app.com/api/webhook)",
						},
						"httpMethod": {
							Type:        "STRING",
							Description: "Método HTTP a ser usado no disparo (GET, POST, PUT, DELETE, PATCH). Padrão é POST.",
						},
						"headers": {
							Type:        "STRING",
							Description: "String JSON válida representando headers HTTP opcionais (ex: '{\"Content-Type\": \"application/json\", \"Authorization\": \"Bearer key\"}')",
						},
						"payload": {
							Type:        "STRING",
							Description: "String JSON válida ou string simples que será enviada como corpo (payload) da requisição para métodos POST/PUT.",
						},
						"webhookAlertUrl": {
							Type:        "STRING",
							Description: "URL opcional para receber alertas caso este job falhe por 3 vezes consecutivas.",
						},
					},
					Required: []string{"name", "schedule", "url"},
				},
			},
			{
				Name:        "triggerJob",
				Description: "Dispara a execução imediata e manual de um Job específico cadastrado utilizando seu ID único.",
				Parameters: GeminiSchema{
					Type: "OBJECT",
					Properties: map[string]GeminiSchema{
						"jobId": {
							Type:        "STRING",
							Description: "O ID da tarefa (formato UUID) obtido na listagem de jobs.",
						},
					},
					Required: []string{"jobId"},
				},
			},
			{
				Name:        "executeCurlDirect",
				Description: "Executa uma requisição HTTP cURL/HTTP imediata e em tempo real, retornando o resultado (status, tempo e corpo da resposta) sem salvar no banco de dados.",
				Parameters: GeminiSchema{
					Type: "OBJECT",
					Properties: map[string]GeminiSchema{
						"url": {
							Type:        "STRING",
							Description: "URL de destino para a requisição.",
						},
						"method": {
							Type:        "STRING",
							Description: "Método HTTP (GET, POST, PUT, DELETE, etc.). Padrão é GET.",
						},
						"headers": {
							Type:        "STRING",
							Description: "JSON contendo cabeçalhos HTTP adicionais (ex: '{\"Authorization\": \"Bearer X\"}').",
						},
						"payload": {
							Type:        "STRING",
							Description: "Corpo da requisição HTTP (JSON ou texto bruto) para métodos POST/PUT.",
						},
					},
					Required: []string{"url"},
				},
			},
		},
	},
}

const systemInstruction = `Você é o CronFlow AI Agent, o assistente inteligente conversacional integrado para automações da plataforma CronFlow.
Seu objetivo é ajudar desenvolvedores e usuários a configurarem e testarem tarefas agendadas em poucos passos e com poucas mensagens.

Regras de Comportamento:
1. Se o usuário fornecer um comando cURL bruto, interprete-o mentalmente, extraia as propriedades necessárias (URL, Headers, Payload, Método HTTP) e configure-o.
2. Se o usuário te der um agendamento informal (ex: "roda toda segunda-feira às 15h"), converta isso para a expressão cron padrão do CronFlow (ex: "0 15 * * 1") ou use o formato simplified "every:Xm/h/d".
3. Seja extremamente prático e conciso. Configure tudo com o mínimo de mensagens possível.
4. Se o usuário pedir para executar ou testar um cURL diretamente, utilize a ferramenta executeCurlDirect para fazer a chamada HTTP instantânea e mostre os resultados.
5. Após criar um job com sucesso, exiba os detalhes formatados e o ID retornado.`

// Chat — POST /v1/agent/chat
// @Summary Conversar com o Agente de IA do CronFlow
// @Description Envia uma mensagem para o Agente de IA, permitindo agendar jobs, testar cURLs, consultar histórico ou gerenciar tarefas por linguagem natural.
// @Tags AI Agent
// @Accept json
// @Produce json
// @Param body body AgentChatRequest true "Histórico e nova mensagem do usuário"
// @Success 200 {object} AgentChatResponse "Resposta do Agente de IA e histórico atualizado"
// @Failure 400 {object} map[string]string "Payload inválido ou sem mensagem"
// @Failure 401 {object} map[string]string "Não autenticado"
// @Failure 500 {object} map[string]string "Erros na API do Gemini/Groq ou durante a execução de ferramentas"
// @Security ApiKeyAuth
// @Router /v1/agent/chat [post]
func (h *AgentHandler) Chat(w http.ResponseWriter, r *http.Request) {
	proj := middleware.ProjectFromContext(r.Context())

	var input AgentChatRequest

	if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
		writeError(w, http.StatusBadRequest, "body invalido")
		return
	}

	if h.cfg.GeminiAPIKey == "" && h.cfg.GroqAPIKey == "" {
		writeError(w, http.StatusInternalServerError, "Chaves de API do Gemini/Groq nao configuradas no backend")
		return
	}

	// Prepare the conversation history
	history := input.History
	
	// Add user's new message to the history
	history = append(history, GeminiMessage{
		Role: "user",
		Parts: []GeminiPart{
			{Text: input.Message},
		},
	})

	ctx := r.Context()
	useGroq := h.cfg.GeminiAPIKey == ""

	if useGroq {
		log.Println("🤖 Iniciando chat com o fallback Groq (LLaMA)")
		reply, updatedHistory, err := h.runGroqChat(ctx, proj.ID, history)
		if err != nil {
			log.Printf("ERROR Groq Chat Execution: %v", err)
			writeError(w, http.StatusInternalServerError, "erro ao processar requisicao com Groq")
			return
		}
		writeJSON(w, http.StatusOK, map[string]any{
			"reply":   reply,
			"history": updatedHistory,
		})
		return
	}

	httpClient := &http.Client{Timeout: 30 * time.Second}
	geminiUrl := "https://generativelanguage.googleapis.com/v1beta/models/gemini-3.5-flash-lite:generateContent?key=" + h.cfg.GeminiAPIKey

	// Execute up to 5 loop iterations to resolve function calls
	for loop := 0; loop < 5; loop++ {
		geminiReq := GeminiRequest{
			Contents: history,
			SystemInstruction: &GeminiSystemInstruction{
				Parts: []GeminiPart{{Text: systemInstruction}},
			},
			Tools: geminiTools,
		}

		reqBytes, err := json.Marshal(geminiReq)
		if err != nil {
			log.Printf("ERROR Gemini Marshal: %v", err)
			writeError(w, http.StatusInternalServerError, "erro ao processar requisicao do agente")
			return
		}

		req, err := http.NewRequestWithContext(ctx, "POST", geminiUrl, bytes.NewReader(reqBytes))
		if err != nil {
			log.Printf("ERROR Gemini Request Creation: %v", err)
			writeError(w, http.StatusInternalServerError, "erro ao criar chamada para IA")
			return
		}
		req.Header.Set("Content-Type", "application/json")

		resp, err := httpClient.Do(req)
		if err != nil || resp.StatusCode != http.StatusOK {
			if err != nil {
				log.Printf("ERROR Gemini API Call: %v. Checking Groq fallback...", err)
			} else {
				var errBody map[string]any
				_ = json.NewDecoder(resp.Body).Decode(&errBody)
				log.Printf("ERROR Gemini API Status %d: %v. Checking Groq fallback...", resp.StatusCode, errBody)
				resp.Body.Close()
			}

			// Fallback to Groq if key exists
			if h.cfg.GroqAPIKey != "" {
				log.Println("🤖 Gemini indisponivel. Ativando fallback Groq (LLaMA)")
				reply, updatedHistory, err := h.runGroqChat(ctx, proj.ID, history)
				if err != nil {
					log.Printf("ERROR Groq Fallback Execution: %v", err)
					writeError(w, http.StatusInternalServerError, "erro ao processar requisicao via Groq fallback")
					return
				}
				writeJSON(w, http.StatusOK, map[string]any{
					"reply":   reply,
					"history": updatedHistory,
				})
				return
			}

			writeError(w, http.StatusInternalServerError, "IA respondeu com erro ou limite atingido")
			return
		}
		defer resp.Body.Close()

		var geminiResp GeminiResponse
		if err := json.NewDecoder(resp.Body).Decode(&geminiResp); err != nil {
			log.Printf("ERROR Gemini Response Decode: %v", err)
			writeError(w, http.StatusInternalServerError, "erro ao decodificar resposta da IA")
			return
		}

		if len(geminiResp.Candidates) == 0 {
			writeError(w, http.StatusInternalServerError, "IA nao retornou nenhuma resposta")
			return
		}

		candidate := geminiResp.Candidates[0]
		
		// Append model's response to history
		history = append(history, candidate.Content)

		hasFunctionCall := false
		var toolParts []GeminiPart

		for _, part := range candidate.Content.Parts {
			if part.FunctionCall != nil {
				hasFunctionCall = true
				result, err := h.executeTool(ctx, proj.ID, part.FunctionCall.Name, part.FunctionCall.Args)
				var responseMap map[string]any
				if err != nil {
					responseMap = map[string]any{"error": err.Error()}
				} else {
					responseMap = map[string]any{"result": result}
				}

				toolParts = append(toolParts, GeminiPart{
					FunctionResponse: &GeminiFunctionResponse{
						ID:       part.FunctionCall.ID,
						Name:     part.FunctionCall.Name,
						Response: responseMap,
					},
				})
			}
		}

		if !hasFunctionCall {
			var reply string
			for _, part := range candidate.Content.Parts {
				if part.Text != "" {
					reply = part.Text
				}
			}
			writeJSON(w, http.StatusOK, map[string]any{
				"reply":   reply,
				"history": history,
			})
			return
		}

		history = append(history, GeminiMessage{
			Role:  "user",
			Parts: toolParts,
		})
	}

	writeError(w, http.StatusInternalServerError, "excedeu o limite de chamadas de funcao do agente")
}

func (h *AgentHandler) executeTool(ctx context.Context, projectID string, name string, args map[string]any) (any, error) {
	switch name {
	case "listJobs":
		jobs, err := h.jobService.List(ctx, projectID)
		if err != nil {
			return nil, err
		}
		return jobs, nil

	case "createJob":
		nameVal, _ := args["name"].(string)
		scheduleVal, _ := args["schedule"].(string)
		urlVal, _ := args["url"].(string)
		if nameVal == "" || scheduleVal == "" || urlVal == "" {
			return nil, fmt.Errorf("name, schedule and url are required")
		}

		httpMethodVal, _ := args["httpMethod"].(string)
		if httpMethodVal == "" {
			httpMethodVal = "POST"
		}

		headersVal := make(map[string]string)
		if headersStr, ok := args["headers"].(string); ok && headersStr != "" {
			_ = json.Unmarshal([]byte(headersStr), &headersVal)
		}

		var payloadVal map[string]any
		if payloadStr, ok := args["payload"].(string); ok && payloadStr != "" {
			_ = json.Unmarshal([]byte(payloadStr), &payloadVal)
			if payloadVal == nil {
				payloadVal = map[string]any{"data": payloadStr}
			}
		}

		var webhookAlertUrl *string
		if alertUrl, ok := args["webhookAlertUrl"].(string); ok && alertUrl != "" {
			webhookAlertUrl = &alertUrl
		}

		created, err := h.jobService.Create(ctx, service.CreateJobInput{
			ProjectID:       projectID,
			Name:            nameVal,
			Schedule:        scheduleVal,
			URL:             urlVal,
			HTTPMethod:      job.HTTPMethod(httpMethodVal),
			Headers:         headersVal,
			Payload:         payloadVal,
			WebhookAlertURL: webhookAlertUrl,
		})
		if err != nil {
			return nil, err
		}
		return created, nil

	case "triggerJob":
		jobID, _ := args["jobId"].(string)
		if jobID == "" {
			return nil, fmt.Errorf("jobId is required")
		}
		err := h.jobService.TriggerNow(ctx, jobID, projectID)
		if err != nil {
			return nil, err
		}
		return map[string]string{"status": "success", "message": "tarefa enfileirada para execucao imediata"}, nil

	case "executeCurlDirect":
		urlVal, _ := args["url"].(string)
		if urlVal == "" {
			return nil, fmt.Errorf("url is required")
		}

		methodVal, _ := args["method"].(string)
		if methodVal == "" {
			methodVal = "GET"
		}

		headersVal := make(map[string]string)
		if headersStr, ok := args["headers"].(string); ok && headersStr != "" {
			_ = json.Unmarshal([]byte(headersStr), &headersVal)
		}

		var payloadBytes []byte
		if payloadStr, ok := args["payload"].(string); ok && payloadStr != "" {
			payloadBytes = []byte(payloadStr)
		}

		// Executa a requisição
		reqClient := &http.Client{Timeout: 15 * time.Second}
		req, err := http.NewRequest(methodVal, urlVal, bytes.NewReader(payloadBytes))
		if err != nil {
			return nil, fmt.Errorf("falha ao criar requisicao: %w", err)
		}

		for k, v := range headersVal {
			req.Header.Set(k, v)
		}
		if len(payloadBytes) > 0 && req.Header.Get("Content-Type") == "" {
			req.Header.Set("Content-Type", "application/json")
		}

		startTime := time.Now()
		resp, err := reqClient.Do(req)
		duration := time.Since(startTime)

		if err != nil {
			return map[string]any{
				"status":  "Error",
				"reason":  err.Error(),
				"latency": fmt.Sprintf("%dms", duration.Milliseconds()),
			}, nil
		}
		defer resp.Body.Close()

		bodyBytes, _ := io.ReadAll(resp.Body)
		bodyStr := string(bodyBytes)
		if len(bodyStr) > 1000 {
			bodyStr = bodyStr[:1000] + "\n...(truncado)"
		}

		respHeaders := make(map[string]string)
		for k, v := range resp.Header {
			if len(v) > 0 {
				respHeaders[k] = v[0]
			}
		}

		return map[string]any{
			"status":  resp.Status,
			"code":    resp.StatusCode,
			"latency": fmt.Sprintf("%dms", duration.Milliseconds()),
			"headers": respHeaders,
			"body":    bodyStr,
		}, nil

	default:
		return nil, fmt.Errorf("tool desconhecida: %s", name)
	}
}

// Groq Fallback logic implementation
func (h *AgentHandler) runGroqChat(ctx context.Context, projectID string, history []GeminiMessage) (string, []GeminiMessage, error) {
	httpClient := &http.Client{Timeout: 30 * time.Second}
	groqUrl := "https://api.groq.com/openai/v1/chat/completions"

	openaiTools := getOpenAITools()
	openaiHistory := translateGeminiToOpenAI(history)

	for loop := 0; loop < 5; loop++ {
		messages := []OpenAIMessage{
			{
				Role:    "system",
				Content: systemInstruction,
			},
		}
		messages = append(messages, openaiHistory...)

		groqReq := OpenAIRequest{
			Model:    "llama-3.3-70b-versatile",
			Messages: messages,
			Tools:    openaiTools,
		}

		reqBytes, err := json.Marshal(groqReq)
		if err != nil {
			return "", nil, err
		}

		req, err := http.NewRequestWithContext(ctx, "POST", groqUrl, bytes.NewReader(reqBytes))
		if err != nil {
			return "", nil, err
		}
		req.Header.Set("Content-Type", "application/json")
		req.Header.Set("Authorization", "Bearer "+h.cfg.GroqAPIKey)

		resp, err := httpClient.Do(req)
		if err != nil {
			return "", nil, err
		}
		defer resp.Body.Close()

		if resp.StatusCode != http.StatusOK {
			var errBody map[string]any
			_ = json.NewDecoder(resp.Body).Decode(&errBody)
			return "", nil, fmt.Errorf("Groq HTTP %d: %v", resp.StatusCode, errBody)
		}

		var groqResp OpenAIResponse
		if err := json.NewDecoder(resp.Body).Decode(&groqResp); err != nil {
			return "", nil, err
		}

		if len(groqResp.Choices) == 0 {
			return "", nil, fmt.Errorf("sem choices do Groq")
		}

		choice := groqResp.Choices[0]
		assistantMsg := choice.Message

		openaiHistory = append(openaiHistory, assistantMsg)

		if len(assistantMsg.ToolCalls) == 0 {
			var finalHistory []GeminiMessage
			for _, m := range openaiHistory {
				finalHistory = append(finalHistory, translateOpenAIToGemini(m))
			}
			return assistantMsg.Content, finalHistory, nil
		}

		// Executa ferramentas para o Groq
		for _, tc := range assistantMsg.ToolCalls {
			var args map[string]any
			_ = json.Unmarshal([]byte(tc.Function.Arguments), &args)

			result, err := h.executeTool(ctx, projectID, tc.Function.Name, args)
			var responseMap map[string]any
			if err != nil {
				responseMap = map[string]any{"error": err.Error()}
			} else {
				responseMap = map[string]any{"result": result}
			}

			respStr, _ := json.Marshal(responseMap)

			openaiHistory = append(openaiHistory, OpenAIMessage{
				Role:       "tool",
				Content:    string(respStr),
				ToolCallID: tc.ID,
			})
		}
	}

	return "", nil, fmt.Errorf("excedeu loop de tools no Groq fallback")
}

func getOpenAITools() []OpenAITool {
	var oTools []OpenAITool
	for _, gt := range geminiTools {
		for _, fd := range gt.FunctionDeclarations {
			oTools = append(oTools, OpenAITool{
				Type: "function",
				Function: OpenAIFunctionDecl{
					Name:        fd.Name,
					Description: fd.Description,
					Parameters:  fd.Parameters,
				},
			})
		}
	}
	return oTools
}

func translateGeminiToOpenAI(history []GeminiMessage) []OpenAIMessage {
	var openaiMsgs []OpenAIMessage
	for _, msg := range history {
		role := msg.Role
		if role == "model" {
			role = "assistant"
		}

		var content string
		var toolCalls []OpenAIToolCall
		var isToolTurn bool
		var toolCallID string

		for _, part := range msg.Parts {
			if part.Text != "" {
				content = part.Text
			} else if part.FunctionCall != nil {
				role = "assistant"
				toolCalls = append(toolCalls, OpenAIToolCall{
					ID:   part.FunctionCall.ID,
					Type: "function",
					Function: OpenAIFunction{
						Name:      part.FunctionCall.Name,
						Arguments: marshalArgs(part.FunctionCall.Args),
					},
				})
			} else if part.FunctionResponse != nil {
				isToolTurn = true
				role = "tool"
				toolCallID = part.FunctionResponse.ID
				content = marshalArgs(part.FunctionResponse.Response)
			}
		}

		if isToolTurn {
			openaiMsgs = append(openaiMsgs, OpenAIMessage{
				Role:       "tool",
				Content:    content,
				ToolCallID: toolCallID,
			})
		} else {
			openaiMsgs = append(openaiMsgs, OpenAIMessage{
				Role:      role,
				Content:   content,
				ToolCalls: toolCalls,
			})
		}
	}
	return openaiMsgs
}

func translateOpenAIToGemini(openaiMsg OpenAIMessage) GeminiMessage {
	role := openaiMsg.Role
	if role == "assistant" {
		role = "model"
	} else if role == "tool" {
		role = "user"
	}

	var parts []GeminiPart
	if openaiMsg.Content != "" {
		parts = append(parts, GeminiPart{Text: openaiMsg.Content})
	}

	for _, tc := range openaiMsg.ToolCalls {
		var args map[string]any
		_ = json.Unmarshal([]byte(tc.Function.Arguments), &args)
		parts = append(parts, GeminiPart{
			FunctionCall: &GeminiFunctionCall{
				ID:   tc.ID,
				Name: tc.Function.Name,
				Args: args,
			},
		})
	}

	if openaiMsg.ToolCallID != "" {
		var resp map[string]any
		_ = json.Unmarshal([]byte(openaiMsg.Content), &resp)
		parts = append(parts, GeminiPart{
			FunctionResponse: &GeminiFunctionResponse{
				ID:       openaiMsg.ToolCallID,
				Name:     openaiMsg.Content,
				Response: resp,
			},
		})
	}

	return GeminiMessage{
		Role:  role,
		Parts: parts,
	}
}

func marshalArgs(args map[string]any) string {
	b, _ := json.Marshal(args)
	return string(b)
}